Well I totally replaced my symbol table and tests still pass.
We now have the symbol table as an immutable HashMap, and store
a (shallow) copy of it in each `TypedExpr` so we can look at any
expr at any time after typechecking and know exactly what's in scope
and such.  So we no longer throw away data that came from typechecking.

I don't remember why I wanted to do this, but it works, and it Feels
Right.  We could store these `ISymtbl`'s in `hir::Expr` instead with
only the nodes that actually introduce new scopes needing to hold on
to a copy of it, but that's actually more work since it's less
consistent, and I'm unconvinced it would consume less memory.  So
here we are.

Miraculously, all unit tests pass.