Ok I need to think about scopes and I'm sleepy.
I need to have a scope that we are updating on the fly as we
gain information about things, but also store an (immutable)
copy of its current state in expressions.

Maybe handle this via a renaming pass in HIR, that generates
a unique name for each var and then we can throw away scope
info?  Nnnnnnoooo because we need somewhere to store that
info for diagnostics.