Ok this is a mess I need to untangle from first principles.
The point of instantiation is that we take a function and replace
its generic parameters with a new type variable.  We make sure that type
variable is the type of the expression being input to the function
param. We do this when the function is called.  When we check this
function, all we need to do is make sure its generic parameters are
matched with each other.

Make it happen.