Update license to LGPL v3
Annoying as the FSF is, you have to look at the Linux kernel and how
successfully they have slowly, patiently bullied hardware manufacturers into
opening specs for various devices, to the benefit of all.  SDL is another
notable success story. Meanwhile BSD and LLVM are delightful but are also,
frankly, the tools of choice for people who want to be assholes to other
people.  Intel's management engine runs BSD, Apple is a primary backer of LLVM,
etc.  So if I'm building infrastructure that is designed to be used by lots
of other people for Serious Stuff, I want its users to be incentivized to not
suck quite so much.

I also considered the Apache license but it's not copyleft as far as I can
tell, it's permissive like the MIT license.  The Mozilla license is copyleft
but frankly fiddly and weak.  So here we are.

This license applies to both the compiler and to the stdlib.  So you can safely
link either with other software and it's Fine, but modifying either for your
particular hardware or whatever requires you to contribute your changes back to
the world.