Working on test_sum2 test failure
I thiiiiiiiiink what is going on is that we have a generic fn that calls
another generic fn, and we detect and monomorphize correctly the first
one but not the second.  So like:

```
fn thing1(i I32, x @A) =
    ...
end

fn thing2(x @A) =
    thing1(3, x)
end

fn main() {} =
	let x = false
    thing2(x)
end
```

thing2 gets specializezd but I don't think thing1 does.  Hmmmmm.
Do we have to explicitly walk the call graph???