oh this monomorph approach won't work either.
We need *both* the source variable and the call site to specialize
a function: the source variable holds the generic function signature
and the call site holds the signature it's actually called with,
because the function source variable has been instantiated in type
checking!  So its type still just contains generic types, and we
need to go to the call site to find the substitution for it!

But that means that to know what function to specialize, we
have to do whole-hog data-flow analysis to trace the call site
back to the corresponding variable(s), and I don't really want to
do that.  Augh!

I WOULD be okay with the previous version's placeholder approach
of only specializing very simple function expressions, even though
I suspect that would explode sooner than I expect.  But... ugh.

...can we just not monomorphize and make everything polymorphic instead?
Just replace all generic params with void pointers?
It's icky, but it should work for now and we can throw it away later, right?

...right?