Make all reformatting integration tests pass
Doesn't mean that the formatting is *good*, just that it's valid.
It's better than it used to be, that's for damn sure, and it's caught
several bugs in the parser and in the type printing code, so that's nice.
Indentation on the reformatter is still kinda cursed, but I guess that's life.