Revert commit 24722db3fc92
Rc'ing types instead of cloning does improve performance of passes and other
tree-manip-stuff by roughly 10%, but is inconvenient enough that it's prolly
not worth it.  Interning them would work better and also be easier since it'd
make all our types Copy.

But that attempt also had a bit of nice code cleanup in it, so I've
tried to preserve it.