Experiment with flattening our tree-based HIR into a vec of nodes.
It's certainly possible, but doing it is at least as much of a pain in the
ass as writing it was in the first place.  And I doubt the benefit
is that high, since we tend to allocate all our IR nodes at once
anyway so locality is probably decentish.  I'm sure the benefit is
PRESENT, but doesn't matter right now.  I just wanted to see what
the API would look like and, as expected, there are some annoying
wrinkles.