Initial pass of removing @ type params.
Pulled in from devel.  I think it will be easier to copy-pasta and rewrite
than to try to merge them.

This just touches the parser and HIR stuff, and now we write the typechecker
bits ourselves.

I think that to start with we will have no inference at all, and then add it in.