We now have the core of closure conversion, get_free_type_params()
Now need to rewrite the rest of it to actually use that power.  However
that works.  Then we can make a version of it that does the same thing
to values and everything should be ezpz lemon squeezy.

Most of the time was spent avoiding stupid boilerplate far too hard,
which converged into the do_to_children() function which actually does
what I want.  Which... I think is actually a visitor.  Hmmm.