Make parser track (partial) origin info.
I don't think I am interested enough in the problem to BRAIN my way through
tracking origin info in the loop part of the pratt parser right now, so it's
just left absent.  That's fine for now.

We might also need to someday add source info for decls, but that'll be easier.